Conditionally include 'alsa/iatomic.h' only when defined DONT_USE_THIS
On Ubuntu 18.04 (maybe other distros) the 'alsa/iatomic.h' is not found.
and the default build fails.

Apparently, the use of atomic stuff in bristol was found to be
somewhat problematic, so it was mostly disabled by default and allowed
only conditionally, when DONT_USE_THIS is defined.

So here we expand this conditional to include of 'alsa/iatomic.h'.

See-also: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=834180