odyssey (Mk3): Assign slider colors similarly to the Mk3 default scheme
The default color scheme has 6 colors (Blue, Green, Yellow, Pink, White,
and Black). The original slider style has a vertical color line on the
face of a slider knob and small vertical colored line at the middle of
the left and right sides. We style it with just a single horizontal line
at the middle of the slider knob face. The color scheme is the default
one:

- OSC1: Blue
- OSC2: Green
- S/H: Yellow
- LFO: Pink
- Noise: White
- Envelopes: Red
- VCF/Others: Black