Filter: Maintain param consistency when changing filter type
When filter type changes (e.g. type 4 -> 0), some related params are
also being manipulated (cutoff and resonance). The general principle
should be to alter the filter operational characteristic, while
maintaining the continuity of the input values from the controls.

This is achieved by tweaking a param's float_val, yet preserving
the param's int_val (set to the input value). This way a care should
be taken in the filter-type (param[4]) change logic, when manipulating
the related params (cutoff and resonance), to recall the param's
float_val from its current int_val and only after that modify the
float_val as needed.

Respectively, the cutoff and resonance params need to properly handle
the changes just as those made in filter-type cases.