odyssey: On skin switching, reset the filter params to current values
This is due to the related filter params (cutoff and resonance) being
manupulated when the filter type change is requested. So after the
skin (and filter) have been changed, we need to reset the controls to
the stored int-values so that the filter's internal characteristics are
recalculated respectively.

This is especially crucial in case when the filter controls were
moved in the alternative skin. This way the altered control values
should be properly carried into the switched skin (and filter).

For Odyssey this affects only filter cutoff and resonance.

See: bristol/filter.c (also filter2.c)