Filter2: Maintain param consistency when changing filter type
Filter2 is a rework of Filter logic; it also supports several filter
types (see param[4]). In pratical use the filter type is fixed per
synth and is set at the synth's init. However, it can be possible to
change the filter type at synth's run-time (if featured). For such a
switch to be effectuated seamlessly, the changes of filter params done
for a specific filter type should be respectively handled per
corresponding params separately.

This is similar to how it's done in `bristol/filter.c`.