odyssey: Use B_FILTER defined macro instead of the hardcoded value
This is to potentially allow a future use of Filter2 (B_FILTER2),
which is a rework of the filter logic. For the most part Filter2 sounds
the same, but type: 0 ("chamblerlain") filter may result in clipping.
Perhaps, gain in "chamberlain" Filter2 needs to be attenuated, as it
was done in Filter.

See also: `bristol/filter.c`