granular: In regrain(), use samplerate from spec.io (needs testing)
Generally, samplerate is being set in spec.io during initialization of the
operator in quantuminit(). The specs should be passed to operate()
function and then to regrain().

Previously, the samplerate was being used from the tentatively internal
variable, which would be set during operator initialization to the supplied
samplerate value. The code has evolved since.

Presently, the correctness of either approach cannot be fully tested, as
the granular synth is not yet operational. Thus needs more testing.